Migration Scrollable/Scrollspy/Popin vers layout.
