+++ src/components/index.ts
@@ -1,6 +1,5 @@
 export {Autocomplete, AutocompleteProps, AutocompleteResult, AutocompleteStyle} from "./autocomplete";
 export {BooleanRadio, BooleanRadioProps, BooleanRadioStyle} from "./boolean-radio";
-export {ButtonBackToTop, ButtonBackToTopProps, ButtonBackToTopStyle} from "./button-back-to-top";
 export {ButtonHelp} from "./button-help";
 export {ButtonMenu, IconMenu, MenuItem} from "./button-menu";
 export {Checkbox, CheckboxProps} from "./checkbox";
@@ -10,16 +9,8 @@ export {Input, InputProps} from "./input";
 export {InputDate, InputDateProps} from "./input-date";
 export {InputTime, InputTimeProps} from "./input-time";
 export {Label, LabelProps, LabelStyle} from "./label";
-export {Panel, PanelButtons, PanelProps, PanelStyle} from "./panel";
-export {Popin, PopinStyle} from "./popin";
-export {Scrollable, ScrollableContext, ScrollableStyle, Sticky} from "./scrollable";
-export {
-    ScrollspyContainer,
-    ScrollspyContainerProps,
-    ScrollspyMenu,
-    ScrollspyMenuProps,
-    ScrollspyStyle
-} from "./scrollspy-container";
+export {Panel, PanelButtons, PanelDescriptor, PanelProps, PanelStyle} from "./panel";
+export {ScrollableContext, ScrollspyContext} from "./scroll-contexts";
 export {Select, SelectProps, SelectStyle} from "./select";
 export {SelectCheckbox, SelectCheckboxProps, SelectCheckboxStyle} from "./select-checkbox";
 export {SelectRadio, SelectRadioProps, SelectRadioStyle} from "./select-radio";
